requirements.txt
pynacl==1.5.0
pyperclip==1.8.2
